bring camera back
Also bringing back the notion of an active argument set in the
data base. You can set that and the db commands no longer
require you to specify any arguments.

camera defers manipulation of the camera object entirely
into the callback. thus it is decoupled from paraview.
the test for the class uses vtk.