add back support to produce arbitrary analyses
Two new classes: Explorers and engine.
Explorers is a mapping between argument sets and actions so that
we can change arguments and have them result in some action.
The specifics of action is abstracted by defering to the engine
class. The vtk_engine module has the first working engine. It
sets up a sample pipeline with a slice filter and changes the slice
filter accordingly before saving out png files.